% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimindex.R
\name{kimindex}
\alias{kimindex}
\title{Calculate Grain Weight from Grain Area (Kim Index) of Bread Wheat}
\usage{
kimindex(data, area_col, remove_na = FALSE)
}
\arguments{
\item{data}{A data frame containing the grain area or grain weight}

\item{area_col}{The name of the column in the data frame that contains the grain area. This name can be any column that holds the grain area data.}
}
\value{
The input data frame with an additional column \code{predicted_gw} representing the predicted individual grain weight from each grain area.
}
\description{
This function predicts the grain weight (mg) for wheat based on the grain area (mm2)
using the equation y = x^1.32, where y is the grain weight and x is the grain area.
}
\examples{
predicted_grain_weight= kim_index(data, "grain_area", remove_na= FALSE OR TRUE)

}
